requirements.txt
# TDLib Python wrapper
python-telegram==0.18.0

# Core dependencies
aiofiles
aiohttp
asyncio
python-dotenv
pyyaml
requests
psutil

# Database
aiosqlite
cryptg

# Media processing
ffmpeg-python
pillow
hachoir
numpy

# Audio streaming
py-tgcalls==0.9.7

# Web scraping and search
beautifulsoup4
yt-dlp
youtube-search
youtube-search-python
spotipy

# Utilities
tgcrypto
unidecode
speedtest-cli
telegraph
wget
pykeyboard
dnspython

# Development and deployment
gitpython
heroku3
